Remove unused files, rename c_simple.rs to c_ast.rs
