+++ src/backend/lang_c/c_simple_ast_builder.rs
@@ -4,27 +4,30 @@
 //! where `rfn` is an instance of `RadecoFunction`, the function returns an instance of
 //! SimpleCAST and we can obtain higher level representation than Radeco IR.
 
-use std::collections::{HashMap, HashSet};
+use super::c_ast;
+use super::c_ast::Ty;
+use super::c_simple_ast::{ValueNode, SimpleCAST, SimpleCASTEdge, ValueEdge, ActionEdge, ActionNode};
 use frontend::radeco_containers::RadecoFunction;
 use middle::ir::{MOpcode, MAddress};
-use middle::ssa::utils;
-use middle::ssa::ssastorage::{NodeData, SSAStorage};
-use middle::ssa::ssa_traits::{SSA, SSAExtra, SSAMod, SSAWalk, ValueInfo};
 use middle::ssa::cfg_traits::CFG;
-use super::c_simple_ast::{ValueNode, SimpleCAST, SimpleCASTEdge, ValueEdge, ActionEdge, ActionNode};
-use super::c_simple;
-use super::c_simple::Ty;
-use petgraph::visit::EdgeRef;
-use petgraph::graph::{Graph, NodeIndex, EdgeIndex, Edges, EdgeReference};
+use middle::ssa::ssa_traits::{SSA, SSAExtra, SSAMod, SSAWalk, ValueInfo};
+use middle::ssa::ssastorage::{NodeData, SSAStorage};
+use middle::ssa::utils;
 use petgraph::{EdgeDirection, Direction, Directed};
+use petgraph::graph::{Graph, NodeIndex, EdgeIndex, Edges, EdgeReference};
+use petgraph::visit::EdgeRef;
+use std::collections::{HashMap, HashSet};
 
 fn is_debug() -> bool {
     cfg!(feature = "trace_log")
 }
 
 /// This constructs SimpleCAST from an instance of RadecoFunction.
-pub fn recover_simple_ast(rfn: &RadecoFunction, fname_map: &HashMap<u64, String>,
-                          strings: &HashMap<u64, String>) -> SimpleCAST {
+pub fn recover_simple_ast(
+    rfn: &RadecoFunction,
+    fname_map: &HashMap<u64, String>,
+    strings: &HashMap<u64, String>,
+) -> SimpleCAST {
     let mut builder = CASTBuilder::new(rfn, fname_map);
     // Recover values
     let data_graph = CASTDataMap::recover_data(rfn, &mut builder.ast, strings);
@@ -95,7 +98,7 @@ impl<'a> CASTBuilder<'a> {
                 }
             }
         }
-        return None
+        return None;
     }
 
     fn args_inorder(&self, call_node: NodeIndex) -> Vec<NodeIndex> {
@@ -103,7 +106,11 @@ impl<'a> CASTBuilder<'a> {
         if self.rfn.callconv.is_none() {
             return Vec::new();
         }
-        let regs_order = self.rfn.callconv.clone().unwrap().args
+        let regs_order = self.rfn
+            .callconv
+            .clone()
+            .unwrap()
+            .args
             .unwrap_or(Vec::new())
             .into_iter()
             .enumerate()
@@ -127,24 +134,35 @@ impl<'a> CASTBuilder<'a> {
         let func_name = {
             if self.datamap.const_nodes.contains(&callee_node) {
                 let addr = self.ssa.constant_value(callee_node).unwrap_or(0);
-                self.fname_map.get(&addr).cloned().unwrap_or("invalid".to_string())
+                self.fname_map.get(&addr).cloned().unwrap_or(
+                    "invalid".to_string(),
+                )
             } else {
                 "unknown".to_string()
             }
         };
         let args = self.args_inorder(call_node)
             .into_iter()
-            .map(|n| self.datamap.var_map.get(&n)
-                 .cloned().unwrap_or(self.ast.unknown))
-             .collect::<Vec<_>>();
+            .map(|n| {
+                self.datamap.var_map.get(&n).cloned().unwrap_or(
+                    self.ast.unknown,
+                )
+            })
+            .collect::<Vec<_>>();
         let ret_val_node = self.return_node(call_node);
-        self.last_action = self.ast.call_func(&func_name, args.as_slice(), self.last_action, ret_val_node);
+        self.last_action = self.ast.call_func(
+            &func_name,
+            args.as_slice(),
+            self.last_action,
+            ret_val_node,
+        );
         self.last_action
     }
 
     fn addr_str(&self, node: NodeIndex) -> String {
-        self.ssa.address(node)
-            .map(|a| format!("{}", a)).unwrap_or("unknown".to_string())
+        self.ssa.address(node).map(|a| format!("{}", a)).unwrap_or(
+            "unknown".to_string(),
+        )
     }
 
     fn recover_action(&mut self, node: NodeIndex) -> NodeIndex {
@@ -153,26 +171,34 @@ impl<'a> CASTBuilder<'a> {
         radeco_trace!("CASTBuilder::recover {:?} @ {:?}", op, node);
         match op {
             MOpcode::OpCall => {
-                // TODO Add proper argument, require prototype from RadecoFunction
                 let ret = self.call_action(node);
                 if is_debug() {
                     let addr = self.addr_str(node);
                     let ops_dbg = self.ssa.operands_of(node);
-                    self.ast.debug_info_at(ret, format!("Call {:?} @ {}", ops_dbg, addr));
+                    self.ast.debug_info_at(
+                        ret,
+                        format!("Call {:?} @ {}", ops_dbg, addr),
+                    );
                 }
                 ret
             }
             MOpcode::OpStore => {
                 let ops = self.ssa.operands_of(node);
-                let dst = self.datamap.var_map.get(&ops[1]).map(|&x| {
-                    self.ast.derefed_node(x).unwrap_or(x)
-                }).unwrap_or(self.ast.unknown);
-                let src = self.datamap.var_map.get(&ops[2])
-                    .cloned().unwrap_or(self.ast.unknown);
+                let dst = self.datamap
+                    .var_map
+                    .get(&ops[1])
+                    .map(|&x| self.ast.derefed_node(x).unwrap_or(x))
+                    .unwrap_or(self.ast.unknown);
+                let src = self.datamap.var_map.get(&ops[2]).cloned().unwrap_or(
+                    self.ast.unknown,
+                );
                 let ret = self.assign(dst, src);
                 if is_debug() {
                     let addr = self.addr_str(node);
-                    self.ast.debug_info_at(ret, format!("*({:?}) = {:?} @ {}", dst, src, addr));
+                    self.ast.debug_info_at(
+                        ret,
+                        format!("*({:?}) = {:?} @ {}", dst, src, addr),
+                    );
                 }
                 ret
             }
@@ -190,8 +216,8 @@ impl<'a> CASTBuilder<'a> {
 
     fn get_block_addr(&self, block: NodeIndex) -> Option<MAddress> {
         match self.ssa.g[block] {
-                NodeData::BasicBlock(addr, _) => Some(addr),
-                _ => None,
+            NodeData::BasicBlock(addr, _) => Some(addr),
+            _ => None,
         }
     }
 
@@ -207,15 +233,21 @@ impl<'a> CASTBuilder<'a> {
     // succ: SSA NodeIndex for destination node
     fn handle_goto(&mut self, ssa_node: NodeIndex, succ: NodeIndex) {
         radeco_trace!("CASTBuilder::handle goto");
-        let ast_node = self.action_map.get(&ssa_node)
-            .cloned().expect("The node should be added to action_map");
-        let succ_node = self.action_map.get(&succ)
-            .cloned().expect("This should not be None");
+        let ast_node = self.action_map.get(&ssa_node).cloned().expect(
+            "The node should be \
+             added to action_map",
+        );
+        let succ_node = self.action_map.get(&succ).cloned().expect(
+            "This should not be None",
+        );
         let label = self.gen_label(succ);
         let goto_node = self.ast.insert_goto_before(ast_node, succ_node, &label);
         if is_debug() {
             let addr = self.addr_str(ssa_node);
-            self.ast.debug_info_at(goto_node, format!("JMP {:?} @ {}", succ_node, addr));
+            self.ast.debug_info_at(
+                goto_node,
+                format!("JMP {:?} @ {}", succ_node, addr),
+            );
         }
     }
 
@@ -224,23 +256,32 @@ impl<'a> CASTBuilder<'a> {
     // true_node: SSA NodeIndex for if-then block
     fn handle_if(&mut self, ssa_node: NodeIndex, selector: NodeIndex, true_node: NodeIndex) {
         radeco_trace!("CASTBuilder::handle_if");
-        let ast_node = self.action_map.get(&ssa_node)
-            .cloned().expect("The node should be added to action_map");
+        let ast_node = self.action_map.get(&ssa_node).cloned().expect(
+            "The node should be \
+             added to action_map",
+        );
         // Add goto statement as `if then` node
         let goto_node = {
-            let dst_node = self.action_map.get(&true_node)
-                .cloned().expect("This should not be None");
+            let dst_node = self.action_map.get(&true_node).cloned().expect(
+                "This should not \
+                 be None",
+            );
             // Edge from `unknown` will be removed later.
             let unknown = self.ast.unknown;
             let label = self.gen_label(true_node);
             self.ast.add_goto(dst_node, &label, unknown)
         };
         // Add condition node to if statement
-        let cond = self.datamap.var_map.get(&selector).cloned().unwrap_or(self.ast.unknown);
+        let cond = self.datamap.var_map.get(&selector).cloned().unwrap_or(
+            self.ast.unknown,
+        );
         let if_node = self.ast.conditional_insert(cond, goto_node, None, ast_node);
         if is_debug() {
             let addr = self.addr_str(ssa_node);
-            self.ast.debug_info_at(goto_node, format!("IF JMP {:?} @ {}", if_node, addr));
+            self.ast.debug_info_at(
+                goto_node,
+                format!("IF JMP {:?} @ {}", if_node, addr),
+            );
         }
     }
 
@@ -264,7 +305,10 @@ impl<'a> CASTBuilder<'a> {
                     if let Some(selector) = self.ssa.selector_in(cur_node) {
                         self.handle_if(cur_node, selector, blk_cond_info.true_side);
                     } else {
-                        radeco_warn!("block with conditional successors has no selector {:?}", cur_node);
+                        radeco_warn!(
+                            "block with conditional successors has no selector {:?}",
+                            cur_node
+                        );
                     }
                 } else {
                     radeco_err!("Unreachable node {:?}", cur_node);
@@ -325,8 +369,11 @@ impl<'a> CASTDataMap<'a> {
     }
 
     // Returns data map from SSAStorage's NodeIndex to SimpleCAST's NodeIndex
-    fn recover_data(rfn: &'a RadecoFunction, ast: &mut SimpleCAST,
-                    strings: &'a HashMap<u64, String>) -> Self {
+    fn recover_data(
+        rfn: &'a RadecoFunction,
+        ast: &mut SimpleCAST,
+        strings: &'a HashMap<u64, String>,
+    ) -> Self {
         let mut s = Self::new(rfn);
         s.prepare_consts(ast, strings);
         s.prepare_regs(ast);
@@ -340,19 +387,34 @@ impl<'a> CASTDataMap<'a> {
         s
     }
 
-    fn handle_binop(&mut self, ret_node: NodeIndex, ops: Vec<NodeIndex>,
-                    expr: c_simple::Expr, ast: &mut SimpleCAST) {
+    fn handle_binop(
+        &mut self,
+        ret_node: NodeIndex,
+        ops: Vec<NodeIndex>,
+        expr: c_ast::Expr,
+        ast: &mut SimpleCAST,
+    ) {
         assert!(ops.len() == 2);
         let ops_mapped = ops.iter()
             .map(|op| self.var_map.get(op).map(|n| *n).unwrap_or(ast.unknown))
             .collect::<Vec<_>>();
         let expr_node = ast.expr(ops_mapped.as_slice(), expr.clone());
-        radeco_trace!("Add {:?} to {:?}, Operator: {:?}", ret_node, expr_node, expr);
+        radeco_trace!(
+            "Add {:?} to {:?}, Operator: {:?}",
+            ret_node,
+            expr_node,
+            expr
+        );
         self.var_map.insert(ret_node, expr_node);
     }
 
-    fn handle_uniop(&mut self, ret_node: NodeIndex, op: NodeIndex,
-                    expr: c_simple::Expr, ast: &mut SimpleCAST) {
+    fn handle_uniop(
+        &mut self,
+        ret_node: NodeIndex,
+        op: NodeIndex,
+        expr: c_ast::Expr,
+        ast: &mut SimpleCAST,
+    ) {
         if let Some(&n) = self.var_map.get(&op) {
             let expr_node = ast.expr(&[n], expr);
             self.var_map.insert(ret_node, expr_node);
@@ -361,8 +423,13 @@ impl<'a> CASTDataMap<'a> {
         }
     }
 
-    fn handle_cast(&mut self, ret_node: NodeIndex, op: NodeIndex,
-                    expr: c_simple::Expr, ast: &mut SimpleCAST) {
+    fn handle_cast(
+        &mut self,
+        ret_node: NodeIndex,
+        op: NodeIndex,
+        expr: c_ast::Expr,
+        ast: &mut SimpleCAST,
+    ) {
         if self.const_nodes.contains(&op) {
             let ast_node = self.var_map.get(&op).cloned().unwrap_or(ast.unknown);
             self.var_map.insert(ret_node, ast_node);
@@ -382,16 +449,15 @@ impl<'a> CASTDataMap<'a> {
         radeco_trace!("CASTBuilder::handle_phi {:?}", node);
         let ops = self.ssa.operands_of(node);
         // Take first available/mappable node of SimpleCAST's node from phi node
-        if let Some(&head) = ops.into_iter()
-           .filter_map(|n| self.var_map.get(&n)).next() {
-           self.var_map.insert(node, head);
+        if let Some(&head) = ops.into_iter().filter_map(|n| self.var_map.get(&n)).next() {
+            self.var_map.insert(node, head);
         }
     }
 
     fn type_from_str(type_str: &str) -> Option<Ty> {
         // TODO More types
         match type_str {
-            "int" => Some(Ty::new(c_simple::BTy::Int, true, 0)),
+            "int" => Some(Ty::new(c_ast::BTy::Int, true, 0)),
             _ => None,
         }
     }
@@ -412,7 +478,10 @@ impl<'a> CASTDataMap<'a> {
         }
         let ops = self.ssa.operands_of(ret_node);
 
-        radeco_trace!("CASTBuilder::update_values opcode: {:?}", self.ssa.opcode(ret_node));
+        radeco_trace!(
+            "CASTBuilder::update_values opcode: {:?}",
+            self.ssa.opcode(ret_node)
+        );
         match self.ssa.opcode(ret_node).unwrap_or(MOpcode::OpInvalid) {
             MOpcode::OpStore => {
                 assert!(ops.len() == 3);
@@ -431,37 +500,40 @@ impl<'a> CASTDataMap<'a> {
                     self.var_map.insert(ret_node, ast_node);
                 }
             }
-            MOpcode::OpAdd => self.handle_binop(ret_node, ops, c_simple::Expr::Add, ast),
-            MOpcode::OpAnd => self.handle_binop(ret_node, ops, c_simple::Expr::And, ast),
-            MOpcode::OpDiv => self.handle_binop(ret_node, ops, c_simple::Expr::Div, ast),
-            MOpcode::OpEq => self.handle_binop(ret_node, ops, c_simple::Expr::Eq, ast),
-            MOpcode::OpGt => self.handle_binop(ret_node, ops, c_simple::Expr::Gt, ast),
+            MOpcode::OpAdd => self.handle_binop(ret_node, ops, c_ast::Expr::Add, ast),
+            MOpcode::OpAnd => self.handle_binop(ret_node, ops, c_ast::Expr::And, ast),
+            MOpcode::OpDiv => self.handle_binop(ret_node, ops, c_ast::Expr::Div, ast),
+            MOpcode::OpEq => self.handle_binop(ret_node, ops, c_ast::Expr::Eq, ast),
+            MOpcode::OpGt => self.handle_binop(ret_node, ops, c_ast::Expr::Gt, ast),
             // XXX Shl might be wrong operator
-            MOpcode::OpLsl => self.handle_binop(ret_node, ops, c_simple::Expr::Shl, ast),
+            MOpcode::OpLsl => self.handle_binop(ret_node, ops, c_ast::Expr::Shl, ast),
             // XXX Shr might be wrong operator
-            MOpcode::OpLsr => self.handle_binop(ret_node, ops, c_simple::Expr::Shr, ast),
-            MOpcode::OpLt => self.handle_binop(ret_node, ops, c_simple::Expr::Lt, ast),
-            MOpcode::OpMod => self.handle_binop(ret_node, ops, c_simple::Expr::Mod, ast),
-            MOpcode::OpMul => self.handle_binop(ret_node, ops, c_simple::Expr::Mul, ast),
+            MOpcode::OpLsr => self.handle_binop(ret_node, ops, c_ast::Expr::Shr, ast),
+            MOpcode::OpLt => self.handle_binop(ret_node, ops, c_ast::Expr::Lt, ast),
+            MOpcode::OpMod => self.handle_binop(ret_node, ops, c_ast::Expr::Mod, ast),
+            MOpcode::OpMul => self.handle_binop(ret_node, ops, c_ast::Expr::Mul, ast),
             // TODO Add `Narrow` info
-            MOpcode::OpNarrow(size) => self.handle_cast(ret_node, ops[0],
-                                                         c_simple::Expr::Cast(size as usize), ast),
-            MOpcode::OpNot => self.handle_uniop(ret_node, ops[0], c_simple::Expr::Not, ast),
-            MOpcode::OpOr => self.handle_binop(ret_node, ops, c_simple::Expr::Or, ast),
+            MOpcode::OpNarrow(size) => {
+                self.handle_cast(ret_node, ops[0], c_ast::Expr::Cast(size as usize), ast)
+            }
+            MOpcode::OpNot => self.handle_uniop(ret_node, ops[0], c_ast::Expr::Not, ast),
+            MOpcode::OpOr => self.handle_binop(ret_node, ops, c_ast::Expr::Or, ast),
             MOpcode::OpRol => unimplemented!(),
             MOpcode::OpRor => unimplemented!(),
             // TODO Add `SignExt`
-            MOpcode::OpSignExt(size) => self.handle_cast(ret_node, ops[0],
-                                                          c_simple::Expr::Cast(size as usize), ast),
-            MOpcode::OpSub => self.handle_binop(ret_node, ops, c_simple::Expr::Sub, ast),
-            MOpcode::OpXor => self.handle_binop(ret_node, ops, c_simple::Expr::Xor, ast),
+            MOpcode::OpSignExt(size) => {
+                self.handle_cast(ret_node, ops[0], c_ast::Expr::Cast(size as usize), ast)
+            }
+            MOpcode::OpSub => self.handle_binop(ret_node, ops, c_ast::Expr::Sub, ast),
+            MOpcode::OpXor => self.handle_binop(ret_node, ops, c_ast::Expr::Xor, ast),
             // TODO Add `ZeroExt`
-            MOpcode::OpZeroExt(size) => self.handle_cast(ret_node, ops[0],
-                                                          c_simple::Expr::Cast(size as usize), ast),
+            MOpcode::OpZeroExt(size) => {
+                self.handle_cast(ret_node, ops[0], c_ast::Expr::Cast(size as usize), ast)
+            }
             MOpcode::OpCall => {
                 self.update_data_graph_by_call(ret_node, ast);
-            },
-            _ => {},
+            }
+            _ => {}
         }
     }
 
@@ -521,8 +593,7 @@ impl<'a> CASTDataMap<'a> {
     }
 }
 
-struct CASTBuilderVerifier {
-}
+struct CASTBuilderVerifier {}
 
 // type Verifier = Fn(NodeIndex, &mut SimpleCAST, &mut CASTDataMap) -> Result<(), String>;
 impl CASTBuilderVerifier {
@@ -534,21 +605,19 @@ impl CASTBuilderVerifier {
         for node in ssa.inorder_walk() {
             match ssa.opcode(node) {
                 Some(MOpcode::OpCall) => {
-                    if let Err(err) =
-                        Self::verify_args_inorder_at(builder, node) {
+                    if let Err(err) = Self::verify_args_inorder_at(builder, node) {
                         errors.push(err);
                     }
-                    if let Err(err) =
-                        Self::verify_call_action_at(builder, node) {
+                    if let Err(err) = Self::verify_call_action_at(builder, node) {
                         errors.push(err);
                     }
-                },
+                }
                 Some(MOpcode::OpStore) => {
                     if let Err(err) = Self::verify_assign_at(builder, node) {
                         errors.push(err);
                     }
-                },
-                _ => {},
+                }
+                _ => {}
             }
         }
         if errors.len() > 0 {
@@ -589,8 +658,7 @@ impl CASTBuilderVerifier {
             return Err("Failed to get dst operand node from SimpleCAST".to_string());
         }
         let assign_node = builder.assign(dst.unwrap(), src.unwrap());
-        let is_err = builder.last_action != assign_node 
-            || !builder.ast.is_assign_node(assign_node);
+        let is_err = builder.last_action != assign_node || !builder.ast.is_assign_node(assign_node);
         if is_err {
             Err("Failed to append assign action".to_string())
         } else {
@@ -599,10 +667,12 @@ impl CASTBuilderVerifier {
     }
 
     // Verify `call_action` made `SimpleCAST::Action(ActionNode::Call(_))` node.
-    fn verify_call_action_at(builder: &mut CASTBuilder, call_node: NodeIndex) -> Result<(), String> {
+    fn verify_call_action_at(
+        builder: &mut CASTBuilder,
+        call_node: NodeIndex,
+    ) -> Result<(), String> {
         let call_node = builder.call_action(call_node);
-        let is_err = builder.last_action != call_node 
-            || !builder.ast.is_call_node(call_node);
+        let is_err = builder.last_action != call_node || !builder.ast.is_call_node(call_node);
         if is_err {
             Err("`CASTBuilder::call_action` is failed.".to_string())
         } else {
@@ -611,24 +681,27 @@ impl CASTBuilderVerifier {
     }
 }
 
-struct CASTDataMapVerifier {
-}
+struct CASTDataMapVerifier {}
 
 type Verifier = Fn(NodeIndex, &mut SimpleCAST, &mut CASTDataMap) -> Result<(), String>;
 impl CASTDataMapVerifier {
     const DELIM: &'static str = "; ";
 
-    fn verify_datamap(datamap: &mut CASTDataMap,
-                      ast: &mut SimpleCAST,
-                      strings: &HashMap<u64, String>) -> Result<(), String> {
+    fn verify_datamap(
+        datamap: &mut CASTDataMap,
+        ast: &mut SimpleCAST,
+        strings: &HashMap<u64, String>,
+    ) -> Result<(), String> {
         Self::verify_prepare(ast, datamap, strings)?;
         Self::verify_ops(ast, datamap)?;
         Ok(())
     }
 
-    fn verify_prepare(ast: &mut SimpleCAST,
-                      datamap: &mut CASTDataMap,
-                      strings: &HashMap<u64, String>) -> Result<(), String> {
+    fn verify_prepare(
+        ast: &mut SimpleCAST,
+        datamap: &mut CASTDataMap,
+        strings: &HashMap<u64, String>,
+    ) -> Result<(), String> {
         datamap.prepare_consts(ast, strings);
         Self::verify_prepare_consts(ast, datamap, strings)?;
         datamap.prepare_regs(ast);
@@ -637,23 +710,38 @@ impl CASTDataMapVerifier {
     }
 
     fn verify_ops(ast: &mut SimpleCAST, datamap: &mut CASTDataMap) -> Result<(), String> {
-        Self::verify_handler_each_node(ast, datamap,
-                                       &Self::verify_handle_uniop, "Handle unary operator")?;
+        Self::verify_handler_each_node(
+            ast,
+            datamap,
+            &Self::verify_handle_uniop,
+            "Handle unary operator",
+        )?;
 
-        Self::verify_handler_each_node(ast, datamap,
-                                       &Self::verify_handle_binop, "Handle binary operator")?;
-        Self::verify_handler_each_node(ast, datamap,
-                                       &Self::verify_handle_cast, "Handle casting operator")?;
+        Self::verify_handler_each_node(
+            ast,
+            datamap,
+            &Self::verify_handle_binop,
+            "Handle binary operator",
+        )?;
+        Self::verify_handler_each_node(
+            ast,
+            datamap,
+            &Self::verify_handle_cast,
+            "Handle casting operator",
+        )?;
         Ok(())
     }
 
-    fn verify_prepare_consts(ast: &SimpleCAST, datamap: &CASTDataMap, strings: &HashMap<u64, String>) -> Result<(), String> {
+    fn verify_prepare_consts(
+        ast: &SimpleCAST,
+        datamap: &CASTDataMap,
+        strings: &HashMap<u64, String>,
+    ) -> Result<(), String> {
         let mut errors = Vec::new();
         // All nodes of datamap.const_nodes should be constant node of SSAStorage
         for &const_node in &datamap.const_nodes {
             if !datamap.ssa.is_constant(const_node) {
-                errors.push(format!("Invalid constant node: {:?}",
-                                    const_node));
+                errors.push(format!("Invalid constant node: {:?}", const_node));
             }
         }
 
@@ -695,15 +783,18 @@ impl CASTDataMapVerifier {
     }
 
     // node: NodeIndex of SSAStorage
-    fn verify_prepare_regs_of(datamap: &CASTDataMap,
-                              node: NodeIndex, name: String) -> Result<(), String> {
+    fn verify_prepare_regs_of(
+        datamap: &CASTDataMap,
+        node: NodeIndex,
+        name: String,
+    ) -> Result<(), String> {
         let mut errors = Vec::new();
-        if  datamap.var_map.get(&node).is_none() {
+        if datamap.var_map.get(&node).is_none() {
             let err = format!("Invalid register node: {:?}", node);
             errors.push(err);
         }
         // Checking if name is a key of reg_map.
-        // reg_map.get(&name) is not needed to be same to 
+        // reg_map.get(&name) is not needed to be same to
         // ast_node of var_map.get(&node)
         if !datamap.reg_map.contains_key(&name) {
             let err = format!("Invalid register name: {:?}", name);
@@ -725,7 +816,12 @@ impl CASTDataMapVerifier {
             }
             let reg_map = utils::register_state_info(reg_state.unwrap(), datamap.ssa);
             for (idx, (node, vt)) in reg_map.into_iter() {
-                let name = datamap.ssa.regfile.get_name(idx).unwrap_or("mem").to_string();
+                let name = datamap
+                    .ssa
+                    .regfile
+                    .get_name(idx)
+                    .unwrap_or("mem")
+                    .to_string();
                 let res = Self::verify_prepare_regs_of(datamap, node, name);
                 if let Err(e) = res {
                     errors.push(e);
@@ -739,8 +835,12 @@ impl CASTDataMapVerifier {
         }
     }
 
-    fn verify_handler_each_node(ast: &mut SimpleCAST, datamap: &mut CASTDataMap,
-                      verifier: &Verifier, name: &str) -> Result<(), String> {
+    fn verify_handler_each_node(
+        ast: &mut SimpleCAST,
+        datamap: &mut CASTDataMap,
+        verifier: &Verifier,
+        name: &str,
+    ) -> Result<(), String> {
         let mut errors = Vec::new();
         for node in datamap.ssa.inorder_walk() {
             if let Err(err) = verifier(node, ast, datamap) {
@@ -748,21 +848,30 @@ impl CASTDataMapVerifier {
             }
         }
         if errors.len() > 0 {
-            Err(format!("{} @ {}", errors.join(Self::DELIM), name.to_string()))
+            Err(format!(
+                "{} @ {}",
+                errors.join(Self::DELIM),
+                name.to_string()
+            ))
         } else {
             Ok(())
         }
     }
 
-    fn verify_handle_uniop(node: NodeIndex, ast: &mut SimpleCAST,
-                           datamap: &mut CASTDataMap) -> Result<(), String> {
+    fn verify_handle_uniop(
+        node: NodeIndex,
+        ast: &mut SimpleCAST,
+        datamap: &mut CASTDataMap,
+    ) -> Result<(), String> {
         // Ensure `handle_uniop` insert node as key into var_map.
-        let expr = c_simple::Expr::Not;
-        let operand_node = datamap.var_map
+        let expr = c_ast::Expr::Not;
+        let operand_node = datamap
+            .var_map
             .iter()
             .map(|(n, _)| *n)
             .filter(|n| *n != node)
-            .next().unwrap();
+            .next()
+            .unwrap();
         // Erase the key so as to ensure whether the key will be correctly inserted
         // by handle_uniop
         datamap.var_map.remove(&node);
@@ -774,15 +883,20 @@ impl CASTDataMapVerifier {
         }
     }
 
-    fn verify_handle_binop(node: NodeIndex, ast: &mut SimpleCAST,
-                           datamap: &mut CASTDataMap) -> Result<(), String> {
+    fn verify_handle_binop(
+        node: NodeIndex,
+        ast: &mut SimpleCAST,
+        datamap: &mut CASTDataMap,
+    ) -> Result<(), String> {
         // Ensure `handle_binop` insert node as key into var_map.
-        let expr = c_simple::Expr::Add;
-        let operand_nodes = datamap.var_map
+        let expr = c_ast::Expr::Add;
+        let operand_nodes = datamap
+            .var_map
             .iter()
             .map(|(n, _)| *n)
             .filter(|n| *n != node)
-            .take(2).collect();
+            .take(2)
+            .collect();
         // Erase the key so as to ensure whether the key will be correctly inserted
         // by handle_binop
         datamap.var_map.remove(&node);
@@ -794,14 +908,21 @@ impl CASTDataMapVerifier {
         }
     }
 
-    fn verify_handle_cast(node: NodeIndex, ast: &mut SimpleCAST, datamap: &mut CASTDataMap) -> Result<(), String> {
+    fn verify_handle_cast(
+        node: NodeIndex,
+        ast: &mut SimpleCAST,
+        datamap: &mut CASTDataMap,
+    ) -> Result<(), String> {
         // Ensure `handle_cast` insert node as key into var_map.
-        let expr = c_simple::Expr::Cast(8);
-        let operand_node = datamap.var_map
+        let expr = c_ast::Expr::Cast(8);
+        let operand_node = datamap
+            .var_map
             .iter()
             .map(|(n, _)| *n)
             .filter(|n| *n != node)
-            .next().unwrap().clone();
+            .next()
+            .unwrap()
+            .clone();
         // Erase the key so as to ensure whether the key will be correctly inserted
         // by handle_uniop
         datamap.var_map.remove(&node);
@@ -816,38 +937,44 @@ impl CASTDataMapVerifier {
 
 #[cfg(test)]
 mod test {
-    use serde_json;
-    use r2api::structs::LRegInfo;
-    use std::sync::Arc;
-    use std::io::prelude::*;
-    use std::fs::File;
-    use std::path::PathBuf;
-    use std::collections::HashMap;
+    use backend::lang_c::{c_ast, c_simple_ast, c_simple_ast_builder};
+    use backend::lang_c::c_simple_ast_builder::{CASTBuilder, CASTDataMap, CASTBuilderVerifier,
+                                                CASTDataMapVerifier};
     use frontend::radeco_containers::RadecoFunction;
     use frontend::radeco_source::SourceErr;
-    use middle::ssa::verifier;
-    use middle::regfile::SubRegisterFile;
     use middle::ir_reader;
-    use backend::lang_c::{c_simple_ast, c_simple_ast_builder};
-    use backend::lang_c::c_simple_ast_builder::{CASTBuilder, CASTDataMap, CASTBuilderVerifier, CASTDataMapVerifier};
+    use middle::regfile::SubRegisterFile;
+    use middle::ssa::verifier;
+    use r2api::structs::LRegInfo;
+    use serde_json;
+    use std::collections::HashMap;
+    use std::fs::File;
+    use std::io::prelude::*;
+    use std::path::PathBuf;
+    use std::sync::Arc;
 
     fn register_profile() -> Result<LRegInfo, SourceErr> {
         let regfile_path = "./test_files/x86_register_profile.json";
         let path = PathBuf::from(regfile_path);
         let mut f = File::open(path).expect("Failed to open file");
         let mut json_str = String::new();
-        let _ = f.read_to_string(&mut json_str).expect("Failed to read file");
+        let _ = f.read_to_string(&mut json_str).expect(
+            "Failed to read file",
+        );
         Ok(serde_json::from_str(&json_str)?)
     }
 
     fn load(name: &str) -> RadecoFunction {
         let ssa = {
-            let regfile = Arc::new(SubRegisterFile::new(&register_profile()
-                    .expect("Unable to load register profile")));
+            let regfile = Arc::new(SubRegisterFile::new(&register_profile().expect(
+                "Unable to load register profile",
+            )));
             let mut f = File::open(name).expect("file not found");
             let mut ir_str = String::new();
-            f.read_to_string(&mut ir_str)
-                .expect("something went wrong reading the file");
+            f.read_to_string(&mut ir_str).expect(
+                "something went wrong reading \
+                 the file",
+            );
             ir_reader::parse_il(&ir_str, regfile)
         };
         let mut rfn = RadecoFunction::default();
@@ -876,8 +1003,12 @@ mod test {
             let mut builder = CASTBuilder::new(&rfn, &dummy_map);
             let data_graph = CASTDataMap::recover_data(&rfn, &mut builder.ast, &dummy_map);
             builder.datamap = data_graph;
-            CASTBuilderVerifier::verify(&mut builder)
-                .expect(&format!("CASTBuilder verification failed {}", file));
+            CASTBuilderVerifier::verify(&mut builder).expect(&format!(
+                "CASTBuilder \
+                 verification \
+                 failed {}",
+                file
+            ));
         }
     }
 }
